requirements.txt
flask==3.1.2
gunicorn==23.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
